bugfix: deeply nested dialogs and inert
When you open an anchored dialog from a modal dialog (e.g. contact card over a conclusion)
the anchored dialog would be inerted after it opened from another place that wasn't a modal
dialog. This fixes that bug by not inerting the dialog after it closes from a modal dialog
unless it was previously inerted.