Add additional margin to the top when in helper or admin mode (#118)
If applied, this commit will add additional margin to the top when in helper or admin mode.

This is for Story S-129160 “Implement new helper/admin mode for mobile responsive sizes”
The story is to restyles the mobile admin and helper mode banner to be a full bar along the top. 

The new style covers more of the modals than it did before. The change in this file detects if the user is currently in helper or admin mode and gives space at the top .

![Screen Shot 2019-07-25 at 5 16 13 PM](https://user-images.githubusercontent.com/6334481/61979888-5b298d00-afb2-11e9-8dff-4f0cacf78198.png)